requirements.txt
# Web 框架
flask>=3.1.0
flask-cors>=5.0.0

# 数据库
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36
flask-sqlalchemy>=3.1.1
pgvector>=0.3.6

# AI 编排
langchain>=0.3.14
langchain-openai>=0.3.0
langchain-community>=0.3.14
langchain-text-splitters>=0.3.0
langgraph>=0.2.60

# Embedding / LLM
openai>=1.59.0

# 文档解析
unstructured>=0.16.11
python-docx>=1.1.2
pypdf>=5.1.0

# 配置管理
python-dotenv>=1.0.1
